Science ideas
The interesting setting you chose for this story would result in some climate and geographical effects that could have important roles in the concepts or design that you choose to explore. I outlined some broad effects that one would be likely to see in such a setting just so you have something to think about. Feel free to ignore these ideas in the name of artistic vision, but I'd be happy to think more about what life on a planet like this would be like.